// js/player/src/generated/TrainingPack.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Guid } from "./Guid.ts";
import type { PlayerId } from "./PlayerId.ts";
import type { Round } from "./Round.ts";

/**
 * Typed training pack model.
 */
export type TrainingPack = { guid: Guid, code: string | null, name: string | null, training_type: string, difficulty: string, creator_name: string | null, description: string | null, tags: Array<number>, map_name: string | null, created_at: bigint, updated_at: bigint, creator_player_id: PlayerId, rounds: Array<Round>, player_team_number: number, unowned: boolean, perfect_completed: boolean, shots_completed: number, };